+++ AdvertisingPlatform.Domain/Services/LocationService.cs
@@ -3,26 +3,29 @@ using AdvertisingPlatforms.Domain.Abstractions;
 using AdvertisingPlatforms.DAL.Entities;
 using AdvertisingPlatforms.DAL.Abstractions;
 using AdvertisingPlatforms.Domain.Models;
+using AdvertisingPlatforms.Base.Exceptions;
+using AdvertisingPlatforms.Base.Constants;
 
 namespace AdvertisingPlatforms.Domain.Services
 {
     public class LocationService : ILocationService
     {
-        private readonly ILogger<LocationService> _logger;
         private readonly IUnitOfWork _unitOfWork;
         private readonly IDomainModelFactory<LocationDb, Location> _factory;
 
-        public LocationService(ILogger<LocationService> logger, 
-            IUnitOfWork unitOfWork,
+        public LocationService(IUnitOfWork unitOfWork,
             IDomainModelFactory<LocationDb, Location> factory)
         {
-            _logger = logger;
             _unitOfWork = unitOfWork;
             _factory = factory;
         }
 
         public async Task<Location> GetById(Guid locationId, CancellationToken cancellationToken)
         {
+            if (!await _unitOfWork.LocationRepository.ExistsAsync(locationId, cancellationToken))
+            {
+                throw new EntityNotFoundExeption(ErrorMessages.ENTITY_NOT_FOUND + locationId);
+            }
             var locationDb = await _unitOfWork.LocationRepository.GetByIdAsync(locationId, cancellationToken);
             var location = _factory.Create(locationDb);
             return location;
@@ -35,26 +38,50 @@ namespace AdvertisingPlatforms.Domain.Services
             return locations;
         }
 
-        public async Task<Location> AddLocation(Location location, CancellationToken cancellationToken)
+        public async Task<Location> CreateLocation(string locationPath, Guid? parentId, CancellationToken cancellationToken)
         {
-            var locationDb = new LocationDb(location.Path, location.ParentId, null);
+            if (await _unitOfWork.LocationRepository.ExistsByPathAsync(locationPath, cancellationToken))
+            {
+                throw new EntityAlreadyExistsExeption(ErrorMessages.ENTITY_ALREADY_EXISTS + locationPath);
+            }
+            if(parentId != null && !await _unitOfWork.LocationRepository.ExistsAsync(parentId.Value, cancellationToken))
+            {
+                throw new EntityNotFoundExeption(ErrorMessages.ENTITY_NOT_FOUND + parentId);
+            }
+            var locationDb = new LocationDb(locationPath, parentId);
             await _unitOfWork.LocationRepository.AddAsync(locationDb, cancellationToken);
             await _unitOfWork.SaveChangesAsync();
             
             return _factory.Create(locationDb);
         }
 
-        public async Task UpdateLocation(Location location, CancellationToken cancellationToken)
+        public async Task UpdateLocation(Guid locationId, string newLocationPath, Guid? newparentId, CancellationToken cancellationToken)
         {
-            var locationDb = await _unitOfWork.LocationRepository.GetByIdAsync(location.Id, cancellationToken);
-            locationDb.Path = location.Path;
-            locationDb.ParentId = location.ParentId;
+            if(!await _unitOfWork.LocationRepository.ExistsAsync(locationId, cancellationToken))
+            {
+                throw new EntityNotFoundExeption(ErrorMessages.ENTITY_NOT_FOUND +  locationId);
+            }
+            if (newparentId != null && !await _unitOfWork.LocationRepository.ExistsAsync(newparentId.Value, cancellationToken))
+            {
+                throw new EntityNotFoundExeption(ErrorMessages.ENTITY_NOT_FOUND + newparentId);
+            }
+            if (await _unitOfWork.LocationRepository.ExistsByPathAsync(newLocationPath, cancellationToken))
+            {
+                throw new EntityAlreadyExistsExeption(ErrorMessages.ENTITY_ALREADY_EXISTS + newLocationPath);
+            }
+            var locationDb = await _unitOfWork.LocationRepository.GetByIdAsync(locationId, cancellationToken);
+            locationDb.Path = newLocationPath;
+            locationDb.ParentId = newparentId;
             await _unitOfWork.LocationRepository.UpdateAsync(locationDb, cancellationToken);
             await _unitOfWork.SaveChangesAsync();
         }
 
         public async Task DeleteLocation(Guid locationId, CancellationToken cancellationToken)
         {
+            if (!await _unitOfWork.LocationRepository.ExistsAsync(locationId, cancellationToken))
+            {
+                throw new EntityNotFoundExeption(ErrorMessages.ENTITY_NOT_FOUND + locationId);
+            }
             await _unitOfWork.LocationRepository.DeleteAsync(locationId, cancellationToken);
             await _unitOfWork.SaveChangesAsync();
         }
